+++ src/cli/cmd/iexec-app.js
@@ -95,6 +95,8 @@ const {
 } = require('../utils/fs');
 const { Keystore } = require('../utils/keystore');
 const { loadChain, connectKeystore } = require('../utils/chains');
+const { lookupAddress } = require('../../common/modules/ens');
+const { ConfigurationError } = require('../../common/utils/errors');
 
 const debug = Debug('iexec:iexec-app');
 
@@ -201,15 +203,40 @@ show
       spinner.start(info.showing(objName));
 
       let res;
+      let ens;
       if (isAddress) {
-        res = await showApp(chain.contracts, addressOrIndex);
+        [res, ens] = await Promise.all([
+          showApp(chain.contracts, addressOrIndex),
+          lookupAddress(chain.contracts, addressOrIndex).catch((e) => {
+            if (e instanceof ConfigurationError) {
+              /** no ENS */
+            } else {
+              throw e;
+            }
+          }),
+        ]);
       } else {
         res = await showUserApp(chain.contracts, addressOrIndex, userAddress);
+        ens = await lookupAddress(chain.contracts, res.objAddress).catch(
+          (e) => {
+            if (e instanceof ConfigurationError) {
+              /** no ENS */
+            } else {
+              throw e;
+            }
+          },
+        );
       }
       const { app, objAddress } = res;
-      spinner.succeed(`App ${objAddress} details:${pretty(app)}`, {
-        raw: { address: objAddress, app },
-      });
+      spinner.succeed(
+        `App ${objAddress} details:${pretty({
+          ...(ens && { ENS: ens }),
+          ...app,
+        })}`,
+        {
+          raw: { address: objAddress, ens, app },
+        },
+      );
     } catch (error) {
       handleError(error, cli, opts);
     }
@@ -273,9 +300,7 @@ publish
           (deployedObj) => deployedObj && deployedObj[chain.id],
         ));
       if (!address) {
-        throw Error(
-          `Missing ${objName}Address and no ${objName} found in "deployed.json" for chain ${chain.id}`,
-        );
+        throw Error(info.missingAddressOrDeployed(objName, chain.id));
       }
       debug('useDeployedObj', useDeployedObj, 'address', address);
       if (useDeployedObj) {
@@ -342,9 +367,7 @@ unpublish
           (deployedObj) => deployedObj && deployedObj[chain.id],
         ));
       if (!address) {
-        throw Error(
-          `Missing ${objName}Address and no ${objName} found in "deployed.json" for chain ${chain.id}`,
-        );
+        throw Error(info.missingAddressOrDeployed(objName, chain.id));
       }
       debug('useDeployedObj', useDeployedObj, 'address', address);
       if (useDeployedObj) {