+++ CHANGELOG.md
@@ -8,6 +8,8 @@ All notable changes to this project will be documented in this file.
 
 - TypeScript interfaces
 - `providerOptions` can be passed to the constructor to configure ethers default provider on bridged chain
+- workerpool API url configuration
+- `show` commands display ENS when configured
 
 ### Changed
 
@@ -15,6 +17,7 @@ All notable changes to this project will be documented in this file.
 - reintegrated `iexec-contracts-js-client`
 - fix a bug transforming an Array into a plain object
 - documentation refactoring
+- fix ENS configure resolution removed useless transaction for EOA
 
 ## [7.0.2] 2022-02-16